+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.20 (build 21118, api 8, 2023-06-15)
+### 1.7.20 (build 21119, api 8, 2023-06-16)
 
 - This seems like a good time for a `refactoring` release in anticipation of
   changes coming in 1.8. Basically this means that a lot of things will be
@@ -338,6 +338,15 @@
   few megs but allows us to log nice full stack traces instead of just
   addresses. Try `print(_babase.native_stack_trace())` if you want to make sure
   its working.
+- (build 21118) Fixed an issue where certain messages such as player-left
+  weren't being send to clients.
+- (build 21118) renamed `bascenev1.screenmessage()` to
+  `bascenev1.broadcastmessage()` to make it more clear that it behaves
+  differently (sending messages to all connected clients instead of just the
+  local screen). There is still a `bascenev1.screenmessage()` but that is now
+  the same local-only version available in babase. Added a temporary warning if
+  calling screenmessage() in a situation that in previous versions would have
+  done a broadcast.
 
 ### 1.7.19 (build 20997, api 7, 2023-01-19)